- changed issue of typing in get_number

+++ ds_behavioral/intent/synthetic_intent_model.py
@@ -240,14 +240,14 @@ class SyntheticIntentModel(AbstractIntentModel):
                 elif at_most == 1:
                     multiplier = np.random.randint(1000, 50000)
                     num = (high - low) if (high - low - counter[index]) < 100000 else counter[index] + multiplier
-                    num_choice = np.linspace(low, high, num=num, endpoint=False)
+                    num_choice = np.linspace(low, high, num=num, dtype=int, endpoint=False)
                     rtn_list = list(np.random.choice(num_choice, size=counter[index], replace=False))
                 elif at_most > 1:
                     section_size = int(counter[index]/at_most) if size % at_most == 0 else int(counter[index]/at_most)+1
                     for _ in range(at_most):
                         multiplier = np.random.randint(1000, 50000)
                         num = (high - low) if (high - low - counter[index]) < 100000 else counter[index] + multiplier
-                        num_choice = np.linspace(low, high, num=num, endpoint=False)
+                        num_choice = np.linspace(low, high, num=num, dtype=int, endpoint=False)
                         rtn_list += list(np.random.choice(num_choice, size=section_size, replace=False))
                     rtn_list = rtn_list[:counter[index]]
                     np.random.shuffle(rtn_list)
@@ -268,14 +268,14 @@ class SyntheticIntentModel(AbstractIntentModel):
                     rtn_list += [low] * counter[index]
                 elif at_most == 1:
                     multiplier = np.random.randint(1000, 50000)
-                    num_choice = np.linspace(low, high, num=counter[index] + multiplier, endpoint=False)
+                    num_choice = np.linspace(low, high, num=counter[index]+multiplier, dtype=float, endpoint=False)
                     rtn_list = list(np.random.choice(num_choice, size=counter[index], replace=False))
                 elif at_most > 1:
                     section_size = int(counter[index] / at_most) if size % at_most == 0 else int(
                         counter[index] / at_most) + 1
                     for _ in range(at_most):
                         multiplier = np.random.randint(1000, 50000)
-                        num_choice = np.linspace(low, high, num=counter[index] + multiplier, endpoint=False)
+                        num_choice = np.linspace(low, high, num=counter[index]+multiplier, dtype=float, endpoint=False)
                         rtn_list += list(np.random.choice(num_choice, size=section_size,  replace=False))
                     rtn_list = rtn_list[:counter[index]]
                     np.random.shuffle(rtn_list)
